+++ config/config.go
@@ -88,6 +88,14 @@ func GetConfig() *Config {
 		panic(err)
 	}
 
+	if cfg.ValidatorMode {
+		cfg.TxIndex.IndexAllTags = false
+		cfg.TxIndex.IndexTags = ""
+
+		cfg.RPC.ListenAddress = ""
+		cfg.RPC.GRPCListenAddress = ""
+	}
+
 	cfg.SetRoot(utils.GetMinterHome())
 	EnsureRoot(utils.GetMinterHome())
 
@@ -223,7 +231,7 @@ type BaseConfig struct {
 	// Address to listen for API connections
 	APIListenAddress string `mapstructure:"api_listen_addr"`
 
-	EnableEvents bool `mapstructure:"enable_events"`
+	ValidatorMode bool `mapstructure:"validator_mode"`
 }
 
 // DefaultBaseConfig returns a default base configuration for a Tendermint node
@@ -243,7 +251,7 @@ func DefaultBaseConfig() BaseConfig {
 		DBPath:            "data",
 		GUIListenAddress:  ":3000",
 		APIListenAddress:  ":8841",
-		EnableEvents:      false,
+		ValidatorMode:     false,
 	}
 }